backend/requirements.txt
# AI Resume Screening Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.19
pydantic==2.10.4

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# NLP
nltk==3.9.1

# ML
scikit-learn==1.6.1
numpy==2.2.2
